Move "standalone" packages to `pkg/`
Keep `bcc` and `elf` in the project root, loosely follow
https://peter.bourgon.org/go-best-practices-2016/#repository-structure
for everything else.
