+++ pkg/tracepipe/trace_pipe.go
