+++ pkg/tracepipe/trace_pipe_test.go
